requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.6
pymssql==2.3.1
python-dotenv==1.0.1
